Debugged. Requires bdrc-DBApps 1.0.4

+++ DRS-BATCH-PROCESSING/delete-batches.sh
@@ -26,12 +26,12 @@ UPLOAD_TRACK=${1?"usage: $ME uploadTrackingFileName BuildListSpec"}
 # get the user and the batch name from the mail errors, look up the build list
 # in the upload track.
 # get the batch path from the build list file
-while IFS='|'  read -ra aLine ; do
+while IFS='|' read -ra aLine; do
 
   user=${aLine[0]}
   batchName=${aLine[1]}
 
-# -w because sometimes users are subsets of other users
+  # -w because sometimes users are subsets of other users
   buildList=$(grep -w "${user}" "$UPLOAD_TRACK" | cut -f2 -d'|' | sort -u)
 
   batchPath=$(grep "${batchName}" "$buildList" | cut -f2 -d'|' | sort -u)
@@ -39,13 +39,17 @@ while IFS='|'  read -ra aLine ; do
   # echo ${user}':'${batchName}':'${buildList}':'${batchPath}':' len is ${#aLine[@]}
 
   printf "Updating DB for %s .." "${batchPath}"
-  updateBuildStatus -d prod:~/.drsBatch.config "$batchPath" FAIL
-  printf "removing path...."
-  if [ -d "$batchPath" ]  ; then
-    rm -rf "$batchPath"
+  update_build_status -D -d prod:~/.drsBatch.config "$batchPath" FAIL
+  rc=$?
+  if [[ $rc == "0" ]]; then
+    printf "removing path...."
+    if [ -d "$batchPath" ]; then
+      rm -rf "$batchPath"
+    else
+      printf " Path %s not found\n" "$batchPath"
+    fi
+    printf "done\n"
   else
-    printf " Path %s not found\n" "$batchPath"
+    printf "update_build_status error %s. Not erasing %s\n" $rc $batchPath
   fi
-  printf "done\n"
 done < <(parseMail.awk $MAIL_TXT | grep -v "${targetErrRE}" $MAIL_DAT | cut -f3,4 -d'|' | sort -u -k1 -t'|')
-
